+++ inet/errors.py
@@ -1,3 +1,7 @@
 
 class ConnectionError(Exception):
     pass
+
+
+class RegisterationError(Exception):
+    pass