+++ tests/endpoint.py
@@ -1,3 +1,0 @@
-from .. import endserver
-
-endserver.start_server()