+++ tests/server.py
@@ -1,11 +1,19 @@
-from .. import server
+import logging
+from inet.inetclient import InetClient
+from inet.server import RoutableServer
+
+logging.basicConfig(level=logging.DEBUG)
+
+proxy = InetClient('tcp://127.0.0.1:3014')
+server = RoutableServer('testservice', proxy, 'tcp://127.0.0.1:4001',
+                        'ipc:///tmp/inet.testservice.sock')
 
 
-def user(req, resp):
-    assert req.data('message') == 'ping'
-    resp.meta('status', 200)
-    resp.data('message', 'pong')
-    resp.send()
+@server.route('hello/world')
+def hello(req, resp):
+    resp.data['greeting'] = 'World'
+    return resp
 
-user_server = server.spawn_server('user', user)
-user_server.join()
+server.setup(localproxy=False)
+server.spawnworkers()
+server.loopforever()